+++ src/View/Helper/ReportingCloud.php
@@ -2,20 +2,20 @@
 declare(strict_types=1);
 
 /**
- * ReportingCloud Zend Framework 3 Module
+ * ReportingCloud Laminas Module
  *
- * Zend Framework 3 Module for ReportingCloud Web API. Authored and supported by Text Control GmbH.
+ * Laminas Module for ReportingCloud Web API. Authored and supported by Text Control GmbH.
  *
  * @link      https://www.reporting.cloud to learn more about ReportingCloud
- * @link      https://git.io/Je5US for the canonical source repository
- * @license   https://github.com/TextControl/txtextcontrol-reportingcloud-php-zf-module/blob/master/LICENSE.md
+ * @link      https://git.io/JexF4 for the canonical source repository
+ * @license   https://git.io/JexFB
  * @copyright © 2020 Text Control GmbH
  */
 
 namespace TxTextControl\ReportingCloud\View\Helper;
 
 use TxTextControl\ReportingCloud\ReportingCloud as TxTextControlReportingCloudReportingCloud;
-use Zend\View\Helper\AbstractHelper;
+use Laminas\View\Helper\AbstractHelper;
 
 /**
  * Class ReportingCloud