+++ test/ServiceManagerFactory.php
@@ -3,8 +3,8 @@ declare(strict_types=1);
 
 namespace TxTextControlTest\ReportingCloud;
 
-use Zend\Mvc\Service\ServiceManagerConfig;
-use Zend\ServiceManager\ServiceManager;
+use Laminas\Mvc\Service\ServiceManagerConfig;
+use Laminas\ServiceManager\ServiceManager;
 
 class ServiceManagerFactory
 {
@@ -33,7 +33,7 @@ class ServiceManagerFactory
     {
         return [
             'modules' => [
-                'Zend\Router',
+                'Laminas\Router',
                 'TxTextControl\ReportingCloud',
             ],